% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{cleanGs3}
\alias{cleanGs3}
\title{Clean output files}
\usage{
cleanGs3(config, config.file, task.id = "GS3")
}
\arguments{
\item{config}{list containing the configuration of the GS3 run to be cleaned}

\item{config.file}{path to the text file to which the configuration for GS3 (corresponding to \code{config}) is written}

\item{task.id}{character containing the task identifier used as prefix for the output files}
}
\description{
Remove the various output files generated by GS3.
}
\author{
Timothee Flutre
}
