% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{crossValFold}
\alias{crossValFold}
\title{Single fold of a cross-validation}
\usage{
crossValFold(task.id, rep.id, fold.ids, fold.id, valid.geno.idx.per.fold, dat,
  col.id, col.trait, binary.trait, genos, config, ped.file, afs, remove.files,
  verbose = 1)
}
\arguments{
\item{task.id}{character containing the task identifier used as prefix for the output files (the fold index will be added)}

\item{rep.id}{identifier of the current replicate}

\item{fold.ids}{vector of fold identifiers}

\item{fold.id}{identifier of the current fold}

\item{valid.geno.idx.per.fold}{vector indicating to which fold a given genotypes belongs (see \code{\link{getPartitionGenos}})}

\item{dat}{data frame with phenotypes}

\item{col.id}{column identifier of the records in \code{dat}}

\item{col.trait}{column identifier of the trait in \code{dat}}

\item{binary.trait}{logical indicating if the trait is binary or not}

\item{genos}{matrix of SNP genotypes}

\item{config}{list containing the generic configuration for GS3}

\item{ped.file}{path to the file containing the pedigree}

\item{afs}{vector of allele frequencies which names are SNP identifiers (column names of \code{genos}); used to compute the variance of additive genotypic values from the variance of additive SNP effects (see Vitezica et al, 2013), then used to compute narrow-sense heritability}

\item{remove.files}{remove files per fold (none/some/all); use \code{"some"} in real-life applications in order to keep estimates of SNP effects per fold, thereby allowing to perform genomic prediction afterwards by averaging them}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
named vector with metrics
}
\description{
Perform a single fold of a K-fold cross-validation with GS3.
}
\seealso{
\code{\link{crossValWithGs3}}
}
\author{
Timothee Flutre
}
