% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{crossValRepWithGs3}
\alias{crossValRepWithGs3}
\title{Replicated cross-validation}
\usage{
crossValRepWithGs3(genos, dat, config, task.id = "GS3",
  binary.trait = FALSE, ped.file = NULL, nb.reps = 50, seed = NULL,
  nb.cores.rep = 1, nb.folds = 10, remove.files = "some",
  nb.cores.fold = 1, cl = NULL, verbose = 1)
}
\arguments{
\item{genos}{matrix of SNP genotypes}

\item{dat}{data frame with phenotypes}

\item{config}{list containing the configuration for GS3}

\item{task.id}{character containing the task identifier used as prefix for the output files (for each fold, its index will be added)}

\item{binary.trait}{logical}

\item{ped.file}{path to the file containing the pedigree}

\item{nb.reps}{number of replicates (a set of folds will be sampled for each replicate)}

\item{seed}{if not NULL, this seed for the pseudo-random number generator will be used to sample as many seeds as the number of replicates, these new seeds being used to shuffle genotypes before partitioning per fold}

\item{nb.cores.rep}{number of cores to launch replicates in parallel (via \code{\link[parallel]{mclapply}}, on Unix-like computers)}

\item{nb.folds}{number of folds}

\item{remove.files}{remove files per fold (none/some/all); use \code{"some"} in real-life applications in order to keep estimates of SNP effects per fold, thereby allowing to perform genomic prediction afterwards by averaging them}

\item{nb.cores.fold}{number of cores to launch folds in parallel (via \code{\link[parallel]{mclapply}}, on Unix-like computers, or \code{\link[parallel]{parLapply}} on Windows); you can also use \code{\link[parallel]{detectCores}}; will be set to 1 if \code{nb.cores.rep} is bigger than 1}

\item{cl}{object returned by \code{\link[parallel]{makeCluster}}, necessary only if \code{nb.cores.fold > 1} and the computer runs Windows; if NULL in such cases, will be created silently}

\item{verbose}{verbosity level (0/1/2); there will be a progress bar only for \code{verbose=1}}
}
\value{
data frame
}
\description{
Perform replicated K-fold cross-validation with GS3, i.e. call \code{\link{crossValWithGs3}} several times, with different seeds.
}
\seealso{
\code{\link{crossValWithGs3}}
}
\author{
Timothee Flutre
}
