% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{execGs3}
\alias{execGs3}
\title{Execute GS3}
\usage{
execGs3(config.file, stdouterr.file = "gs3_stdouterr.txt")
}
\arguments{
\item{config.file}{path to the text file containing the configuration for GS3}

\item{stdouterr.file}{path to the text file to which the stdout and stderr will be written}
}
\value{
return value (0 if success)
}
\description{
Execute GS3 via a system call.
}
\author{
Timothee Flutre
}

