% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{getDefaultConfig}
\alias{getDefaultConfig}
\title{Default configuration}
\usage{
getDefaultConfig(nb.snps = NA, method = "VCE", ptl = NULL, twc = NA,
  rec.id = NA)
}
\arguments{
\item{nb.snps}{number of SNPs}

\item{method}{BLUP/MCMCBLUP/VCE/PREDICT}

\item{ptl}{data.frame indicating, for each effect, its position in the data file, its type, and its number of levels}

\item{twc}{vector}

\item{rec.id}{numeric}
}
\value{
list
}
\description{
Return a list corresponding to a default configuration as used in the vignette.
}
\seealso{
\code{\link{writeConfigForGs3}}, \code{\link{isValidConfig}}
}
\author{
Timothee Flutre
}
