% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{getPartitionGenos}
\alias{getPartitionGenos}
\title{Partition for cross-validation}
\usage{
getPartitionGenos(geno.names, nb.folds = 10, seed = NULL)
}
\arguments{
\item{geno.names}{vector of genotype names}

\item{nb.folds}{number of folds}

\item{seed}{if not NULL, this seed for the pseudo-random number generator will be used to shuffle genotypes before partitioning per fold}
}
\value{
vector
}
\description{
Return a vector which content corresponds to fold indices and names to genotypes.
The fold index of a given genotype indicates that, for this fold, the genotype won't be used for training but for validation.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{crossValWithGs3}}
}

