% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{isValidConfig}
\alias{isValidConfig}
\title{Check configuration}
\usage{
isValidConfig(config)
}
\arguments{
\item{config}{list containing the configuration for GS3}
}
\value{
logical
}
\description{
Check configuration
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{getDefaultConfig}}, \code{\link{writeConfigForGs3}}
}

