% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{isValidData}
\alias{isValidData}
\title{Check data}
\usage{
isValidData(x, inds, col.id, col.traits, binary.traits)
}
\arguments{
\item{x}{data frame}

\item{inds}{vector}

\item{col.id}{numeric}

\item{col.traits}{numeric vector}

\item{binary.traits}{logical vector}
}
\value{
logical
}
\description{
Return TRUE if the inputs are properly formatted and consistent.
}
\author{
Timothee Flutre
}
