% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{isValidGenos}
\alias{isValidGenos}
\title{Check genotypes}
\usage{
isValidGenos(x, inds)
}
\arguments{
\item{x}{matrix with SNP genotypes}

\item{inds}{vector}
}
\value{
logical
}
\description{
Return TRUE if the input corresponds to a properly-formatted genotype matrix.
}
\author{
Timothee Flutre
}
