% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs3.R
\name{writeDataForGs3}
\alias{writeDataForGs3}
\title{Write data}
\usage{
writeDataForGs3(x, file, inds, col.id, col.trait, binary.trait = FALSE)
}
\arguments{
\item{x}{data.frame containing trait values, covariables, cross-classified factors and individual identifiers (names of parameter "inds")}

\item{file}{path to the text file to which x will be written}

\item{inds}{named vector with values from 1 to I and names the corresponding individual identifiers}

\item{col.id}{column index of the individual identifiers in x}

\item{col.trait}{column index of the trait in x}

\item{binary.trait}{boolean indicating if the trait is binary or not}
}
\value{
nothing
}
\description{
Write the data file for GS3.
}
\author{
Timothee Flutre
}

