% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{writeGenosForGs3}
\alias{writeGenosForGs3}
\title{Write genotypes}
\usage{
writeGenosForGs3(x, file, inds)
}
\arguments{
\item{x}{matrix with SNP genotypes encoded as allele dose (i.e. 0/1/2 or NA/5), with individuals in rows and SNPs in columns. Row names are compulsory. Missing data coded as "NA" will be replaced by "5". Missing data already coded as "5" won't be modified.}

\item{file}{path to the text file to which x will be written}

\item{inds}{named vector with values from 1 to I and names the corresponding individual identifiers}
}
\value{
nothing
}
\description{
Write the genotype file for GS3.
}
\author{
Timothee Flutre
}

