% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{crossValWithGs3}
\alias{crossValWithGs3}
\title{Cross-validation}
\usage{
crossValWithGs3(genos, dat, col.id = 1, col.trait = 2,
  binary.trait = FALSE, ped.file = NULL, config, nb.folds = 10,
  seed = NULL, verbose = 1)
}
\arguments{
\item{genos}{matrix of SNP genotypes}

\item{dat}{data.frame with phenotypes}

\item{col.id}{column index of the individual identifiers in \code{dat}}

\item{col.trait}{column index of the trait of interest in \code{dat}}

\item{binary.trait}{logical}

\item{ped.file}{path to the file containing the pedigree}

\item{config}{list}

\item{nb.folds}{number of folds}

\item{seed}{if not NULL, this seed for the pseudo-random number generator will be used to shuffle genotypes before partitioning per fold}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
data.frame
}
\description{
Perform K-fold cross-validation with GS3, and report metrics as advised in \href{http://dx.doi.org/10.1534/genetics.112.147983}{Daetwyler et al. (2013)}.
}
\author{
Helene Muranty, Timothee Flutre
}
\seealso{
\code{\link{execGs3}}, \code{\link{getPartitionGenos}}
}

