% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{execGs3}
\alias{execGs3}
\title{Execute GS3}
\usage{
execGs3(config.file, task.id = "GS3")
}
\arguments{
\item{config.file}{path to the text file containing the configuration for GS3}

\item{task.id}{character containing the task identifier used as prefix for the "stdout/stderr" and "freq" output files}
}
\value{
path to the text file to which the stdout and stderr will be written
}
\description{
Execute GS3 via a system call.
The expected executable name in the PATH is \code{gs3.exe} on Windows, \code{gs3} otherwise.
}
\seealso{
\code{\link{writeConfigForGs3}}, \code{\link{vcs2mcmc}}, \code{\link{cleanGs3}}
}
\author{
Timothee Flutre
}
