% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{vcs2mcmc}
\alias{vcs2mcmc}
\title{Load GS3 results}
\usage{
vcs2mcmc(vcs.file)
}
\arguments{
\item{vcs.file}{path to the file containing the variance components' samples}
}
\value{
\code{\link[coda]{mcmc.list}}
}
\description{
Read the file containing the variance components' samples into a \code{\link[coda]{mcmc.list}} object.
}
\examples{
\dontrun{vcs <- vcs2mcmc(vcs.file)
summary(vcs)
coda::effectiveSize(vcs)
genos <- as.matrix(read.table(genos.file))
afs <- colMeans(genos) / 2
tmp <- cbind(as.matrix(vcs[[1]]), varA=vcs[[1]][,"vara"] * 2 * sum(afs * (1 - afs)))
vcs[[1]] <- coda::as.mcmc(tmp)
summary(vcs)}
}
\seealso{
\code{link{execGs3}}
}
\author{
Timothee Flutre
}
