% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{writeDataForGs3}
\alias{writeDataForGs3}
\title{Write data}
\usage{
writeDataForGs3(x, file, inds, col.id = 1, col.traits = 2,
  binary.traits = FALSE)
}
\arguments{
\item{x}{data.frame containing individual identifiers (names of parameter "inds"), trait values (possibly several traits), covariables and cross-classified factors. Missing data coded as "NA" will be replaced by "0" if it is in a 'trait' column and the trait is binary, and by "-9999" otherwise.}

\item{file}{path to the text file to which x will be written}

\item{inds}{named vector with values from 1 to I and names the corresponding individual identifiers}

\item{col.id}{column index of the individual identifiers in x}

\item{col.traits}{column indices of traits in x}

\item{binary.traits}{logical vector indicating if traits are binary or not (should be of the same length as col.traits)}
}
\value{
nothing
}
\description{
Write the data file for GS3.
}
\author{
Timothee Flutre
}

